change architecture + add  data layer and move model
